Split the admin auth test into login, logout, and session suites
